add tests, validations, update readme and create new interfaces

+++ src/main/java/org/edu/fabs/exchangerate/service/impl/ProductServiceImpl.java
@@ -2,10 +2,12 @@ package org.edu.fabs.exchangerate.service.impl;
 
 import com.google.gson.Gson;
 import feign.FeignException;
+import org.apache.commons.lang3.EnumUtils;
+import org.edu.fabs.exchangerate.dto.ExchangeRateResponse;
 import org.edu.fabs.exchangerate.feign.ExchangeFeignClient;
-import org.edu.fabs.exchangerate.handler.BusinessException;
+import org.edu.fabs.exchangerate.handler.InvalidCurrencyCodeException;
+import org.edu.fabs.exchangerate.handler.ResourceNotFoundException;
 import org.edu.fabs.exchangerate.model.CurrencySymbol;
-import org.edu.fabs.exchangerate.model.ExchangeRateResponse;
 import org.edu.fabs.exchangerate.model.Product;
 import org.edu.fabs.exchangerate.repository.ProductRepository;
 import org.edu.fabs.exchangerate.service.ProductService;
@@ -14,6 +16,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -37,7 +40,7 @@ public class ProductServiceImpl implements ProductService {
     @Transactional(readOnly = true)
     @Override
     public Optional<Product> getById(Long id) {
-        return productRepository.findById(id);
+        return Optional.ofNullable(productRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(id)));
     }
 
     @Transactional
@@ -50,7 +53,10 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public Product updateProduct(Long id, Product productToUpdate) {
 
-        Optional<Product> optionalProduct = this.productRepository.findById(id);
+        Optional<Product> optionalProduct = Optional.ofNullable(this.productRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(id)));
+
+        validateCurrencySymbol(productToUpdate.getCurrency());
 
         if (optionalProduct.isPresent()) {
             Product productDB = optionalProduct.get();
@@ -59,27 +65,45 @@ public class ProductServiceImpl implements ProductService {
             productDB.setCurrency(productToUpdate.getCurrency());
             return this.productRepository.save(productDB);
         } else {
-            throw new BusinessException("product not found");
+            throw new NoSuchElementException("Product not found " + productToUpdate);
         }
     }
 
     @Transactional
     @Override
     public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
+        Optional<Product> product = productRepository.findById(id);
+
+        if (product.isPresent()) {
+            productRepository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException(id);
+        }
     }
 
     public BigDecimal calculateTotalPrice(Product product, CurrencySymbol targetCurrency) {
-        BigDecimal conversionRate = BigDecimal.ZERO;
-        try {
-            String response = exchangeFeignClient.getPairConversion(product.getCurrency(), targetCurrency);
-            Gson gson = new Gson();
-            ExchangeRateResponse exchangeRateResponse = gson.fromJson(response, ExchangeRateResponse.class);
-            conversionRate = exchangeRateResponse.getConversion_rate();
-        } catch (FeignException e) {
-            throw new BusinessException("failed to get exchange rate");
-        }
+        validateCurrencySymbol(targetCurrency);
+
+        ExchangeRateResponse exchangeRateResponse = fetchExchangeRate(product.getCurrency(), targetCurrency);
+        BigDecimal conversionRate = exchangeRateResponse.getConversion_rate();
+
         return product.getPrice().multiply(new BigDecimal(product.getQuantity())).multiply(conversionRate);
+    }
+
+    private void validateCurrencySymbol(CurrencySymbol targetCurrency) {
+        if (!EnumUtils.isValidEnum(CurrencySymbol.class, targetCurrency.getName())) {
+            throw new InvalidCurrencyCodeException("Invalid currency code. Currency code must be a valid ISO 4217 code: ", targetCurrency.getName());
+        }
+    }
+
+    public ExchangeRateResponse fetchExchangeRate(CurrencySymbol baseCurrency, CurrencySymbol targetCurrency) {
+        try {
+            String response = exchangeFeignClient.getPairConversion(baseCurrency, targetCurrency);
+            Gson gson = new Gson();
+            return gson.fromJson(response, ExchangeRateResponse.class);
+        } catch (FeignException e) {
+            throw new InvalidCurrencyCodeException("Conversion rate not found for currency pair: ", baseCurrency.getName() + ", " + targetCurrency.getName());
+        }
     }
 
 }
